# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "KAVACH \u2014 AI-Powered DevSecOps Platform"
  description: "KAVACH scans banking application repositories for vulnerabilities,\
    \ calculates a Banking Risk Score (BRS), maps findings to regulatory controls\
    \ (RBI / PCI-DSS / SWIFT), and generates AI-powered audit reports.\n\n## Pipeline\n\
    `GitHub webhook` \u2192 `Create Scan Job` \u2192 `Queue` \u2192 `Distributed workers`\
    \ \u2192 `Aggregation` \u2192 `Risk Engine` \u2192 `Compliance Engine` \u2192\
    \ `AI Explanation` \u2192 `Reports` \u2192 `Dashboard` \u2192 `Notifications`\
    \ \u2192 `Archive`\n\nA scan enters this pipeline either via `POST /scan/repository`\
    \ / `POST /scan` (an authenticated user submitting a URL or archive) or `POST\
    \ /webhooks/github` (an unauthenticated, HMAC-verified push event) \u2014 both\
    \ converge on the same queue \u2192 distributed-worker \u2192 aggregation path,\
    \ so everything downstream behaves identically regardless of how the scan was\
    \ triggered."
  contact:
    name: KAVACH Platform Team
    url: https://github.com/kavach-platform/kavach
  license:
    name: "Proprietary \u2014 All Rights Reserved"
  version: 1.0.0
paths:
  /api/v1/repositories:
    get:
      tags:
      - Repositories
      summary: List Repositories
      description: List the current user's submitted repositories (upload targets
        and repo-URL submissions alike).
      operationId: list_repositories_api_v1_repositories_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RepositoryResponse'
                title: Response List Repositories Api V1 Repositories Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/repositories/{repository_id}/scheduled-scan:
    patch:
      tags:
      - Repositories
      summary: Update Scheduled Scan
      description: "Opt a repository in to (or out of) nightly re-scanning \u2014\
        \ see\napp/tasks/scheduled_scan_tasks.py. Only URL-based repositories can\
        \ be\nscheduled: a one-time zip upload has no re-fetchable source for a\n\
        later re-scan to run against."
      operationId: update_scheduled_scan_api_v1_repositories__repository_id__scheduled_scan_patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Repository Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledScanUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/{scan_job_id}:
    get:
      tags:
      - Reports
      summary: Get Report Status
      description: Check which reports are ready for download for a given scan job,
        and the status of any still in progress.
      operationId: get_report_status_api_v1_reports__scan_job_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: scan_job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Scan Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPathsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/{scan_job_id}/download/{report_type}:
    get:
      tags:
      - Reports
      summary: Download Report
      description: 'Download a specific generated report. For the "s3" storage backend

        this redirects (307) to a short-lived presigned URL rather than

        streaming the file through this process; for "local" it streams the

        file directly, unchanged from before S3 support existed.'
      operationId: download_report_api_v1_reports__scan_job_id__download__report_type__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: scan_job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Scan Job Id
      - name: report_type
        in: path
        required: true
        schema:
          type: string
          title: Report Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RepoProviderType:
      type: string
      enum:
      - upload
      - github
      - gitlab
      - bitbucket
      title: RepoProviderType
    ReportPathsResponse:
      properties:
        scan_job_id:
          type: string
          format: uuid
          title: Scan Job Id
        pdf_available:
          type: boolean
          title: Pdf Available
        pdf_technical_available:
          type: boolean
          title: Pdf Technical Available
        sarif_available:
          type: boolean
          title: Sarif Available
        sbom_available:
          type: boolean
          title: Sbom Available
        unified_findings_available:
          type: boolean
          title: Unified Findings Available
        compliance_report_available:
          type: boolean
          title: Compliance Report Available
        csv_available:
          type: boolean
          title: Csv Available
        reports:
          items:
            $ref: '#/components/schemas/ReportStatusDetail'
          type: array
          title: Reports
          default: []
      type: object
      required:
      - scan_job_id
      - pdf_available
      - pdf_technical_available
      - sarif_available
      - sbom_available
      - unified_findings_available
      - compliance_report_available
      - csv_available
      title: ReportPathsResponse
    ReportStatusDetail:
      properties:
        report_type:
          type: string
          title: Report Type
        status:
          type: string
          title: Status
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - report_type
      - status
      title: ReportStatusDetail
      description: "Per-report-type detail \u2014 lets a client show \"generating\"\
        /\"failed\" instead of just a missing boolean."
    RepositoryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        provider:
          $ref: '#/components/schemas/RepoProviderType'
        default_branch:
          anyOf:
          - type: string
          - type: 'null'
          title: Default Branch
        scheduled_scan_enabled:
          type: boolean
          title: Scheduled Scan Enabled
          default: false
      type: object
      required:
      - id
      - name
      - provider
      title: RepositoryResponse
...
    ScheduledScanUpdateRequest:
      properties:
        enabled:
          type: boolean
          title: Enabled
      type: object
      required:
      - enabled
      title: ScheduledScanUpdateRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...